Fixed app crash, made thrust on left stick sticky, made gyro amplification a float (starting 0,5)

+++ src/se/bitcraze/crazyfliecontrol/controller/Controls.java
@@ -72,7 +72,7 @@ public class Controls {
 
     //Gyro sensor values
     private boolean mUseGyro;
-    private int mGyroAmplification;
+    private float mGyroAmplification;
     private String mGyroAmplificationDefaultValue;
     private boolean mTouchThrustFullTravel;
 
@@ -104,7 +104,7 @@ public class Controls {
 
         mTrimDefaultValue = res.getString(R.string.preferences_trim_defaultValue);
 
-        mGyroAmplificationDefaultValue = Integer.toString(R.string.preferences_gyro_amp_defaultValue);
+        mGyroAmplificationDefaultValue = Float.toString(R.string.preferences_gyro_amp_defaultValue);
 
         mControllerTypeDefaultValue = res.getString(R.string.preferences_controller_defaultValue);
 
@@ -129,7 +129,7 @@ public class Controls {
         this.mControllerType = Integer.parseInt(mPreferences.getString(PreferencesActivity.KEY_PREF_CONTROLLER, mControllerTypeDefaultValue));
 
         this.mUseGyro = mPreferences.getBoolean(PreferencesActivity.KEY_PREF_USE_GYRO_BOOL, false);
-        this.mGyroAmplification = Integer.parseInt(mPreferences.getString(PreferencesActivity.KEY_PREF_GYRO_AMP, mGyroAmplificationDefaultValue));
+        this.mGyroAmplification = Float.parseFloat(mPreferences.getString(PreferencesActivity.KEY_PREF_GYRO_AMP, mGyroAmplificationDefaultValue));
 
         this.mTouchThrustFullTravel = mPreferences.getBoolean(PreferencesActivity.KEY_PREF_TOUCH_THRUST_FULL_TRAVEL, true);
 
@@ -241,7 +241,7 @@ public class Controls {
         return mUseGyro;
     }
 
-    public int getGyroAmplification() {
+    public float getGyroAmplification() {
         return mGyroAmplification;
     }
